fixed navigate after login and signup
